Move chooser activity blacklist to Settings/Apps & Notifications [1/2]
The miscellaneous section is not meant to be a permanent place for tweaks.
Once we figure out where a better place for tweaks, we move them there. This
is what's happening now.

This tweak fits better under Apps & Notifications where things like permissions,
notifications and other options relating to apps are handled.

- Added the 'add' icon to the preference to match similar preferences (Heads up)

Change-Id: Ib70a3a1c97e7a5f0fc36b2be0a3eb8db3c96a247
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>