DUI settings custom action dialogs: fix scroll lag
to replicate: try to set a custom action for a smartbar
button or a fling gesture, after the actions dialog appears
start to scroll down the list.
At least the first scroll is usually glitchy.

We can avoid the issue loading actions and related icons/labels
as soon as we show the dialog.

Change-Id: I07d013442d8806a7dc7e4071b821c7e20ce27fbb
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>